Add chart hover tooltips, time filter dropdown, sidebar UX & dark title bar
- Chart hover: show nearest data point marker and tooltip (time + value)
  on both the cumulative RP and rolling RP/h charts
- Time filter: replace three checkboxes with a ComboBox (All time, 1 week,
  48h, 24h, 12h, 6h, 3h, 2h, 1h, Custom); Custom opens a popup with hours
  and minutes inputs; changing filter auto-restarts an active watch session
- Sidebar toggle: move collapse button from toolbar to the Summary header
  (matching Absolute Numbers / Absolute RPs / Percentages); add a ▶ strip
  at the left edge so the sidebar can be reopened when hidden

+++ DAoCLogWatcher.Core/LogWatcher.cs
@@ -19,7 +19,7 @@ public sealed partial class LogWatcher: IDisposable, IAsyncDisposable
 	private static readonly Regex KillLineRegex = GenerateKillLineRegex();
 
 	private readonly string logFilePath;
-	private readonly int maxHistoryHours;
+	private readonly double maxHistoryHours;
 	private readonly byte[] readBuffer;
 	private FileStream? fileStream;
 	private readonly StringBuilder incompleteLineBuffer;
@@ -28,7 +28,7 @@ public sealed partial class LogWatcher: IDisposable, IAsyncDisposable
 	private readonly Dictionary<string, int> characterNameCounts = new();
 	private bool lastTimestampedLineWasInWindow = true;
 
-	public LogWatcher(string logFilePath, long startPosition = 0, bool enableTimeFiltering = false, int filterHours = 24)
+	public LogWatcher(string logFilePath, long startPosition = 0, bool enableTimeFiltering = false, double filterHours = 24)
 	{
 		ArgumentException.ThrowIfNullOrWhiteSpace(logFilePath);
